Registration Error DOES NOT insert into DB
Put the insert into db code inside of an 'else' statement so that it doesn't automatically insert bad data even though a user generates an error'. Everything should be working as far as the DB goes but need to now connect this to our game and to make sure registered user is able to log in.